Client-提交11.5.2实现群聊天的逻辑以及成员显示与添加
1.群组成员聊天的逻辑实现
2.群成员列表显示与添加的功能实现
3.引入依赖包实现底部面板切换功能
4.修复服务器端创建群的推送逻辑 实现添加群后成员自动刷新群列表

+++ App/ITalker/factory/src/main/java/com/example/ggxiaozhi/factory/presenter/message/ChatContract.java
@@ -57,6 +57,6 @@ public interface ChatContract {
         void showAdminOption(boolean isAdmin);
 
         //初始化群成员信息
-        void onInitGroupMembers(List<MemberUserModel> memberUserModels, int memberCount);
+        void onInitGroupMembers(List<MemberUserModel> memberUserModels, long memberCount);
     }
 }
